Refactor network handling by relocating DirectionSerializer and NetworkTickHandler
