+++ src/main/java/com/logistics/pipe/modules/RequesterModule.java
@@ -4,7 +4,6 @@ import com.logistics.LogisticsPipe;
 import com.logistics.pipe.network.FulfillmentMode;
 import com.logistics.pipe.network.ILogisticsNetwork;
 import com.logistics.core.lib.resource.ResourceId;
-import com.logistics.core.lib.storage.DirectionSerializer;
 import com.logistics.core.lib.storage.NbtCompat;
 import com.logistics.pipe.PipeContext;
 import com.logistics.pipe.block.entity.PipeBlockEntity;
